[Package] [logmonitor] Cleanup Code Base & Move Constant
